+++ src/index.ts
@@ -10,11 +10,18 @@ export {
   EthereumDeployContractPayload,
   Peer,
   SwapRequest,
-  SwapDetails
+  SwapDetails,
+  HalightLightningBitcoinHanEthereumEtherRequestBody,
+  HalightLightningBitcoinHerc20EthereumErc20RequestBody,
+  HanEthereumEtherHalightLightningBitcoinRequestBody,
+  Herc20EthereumErc20HalightLightningBitcoinRequestBody,
+  HalightLightningBitcoinRequestParams,
+  HanEthereumEtherRequestParams,
+  Herc20EthereumErc20RequestParams
 } from "./cnd/cnd";
 export { Problem } from "./cnd/axios_rfc7807_middleware";
 export * from "./cnd/siren";
-export { Transaction, SwapTransactionStatus } from "./transaction";
+export { Transaction, TransactionStatus } from "./transaction";
 
 export { Actor, createActor } from "./actor";